feat(user)/#54 : 가게 선택 페이지 구현
- Hero 애니메이션 메서드 활용
   - tag 속성을 사용하여 tag 이름을 지정하고 비교 후 일치하면 페이지 이동되도록 구현
